# docs/openapi.yaml
openapi: 3.0.3
info:
  title: さて、羊に戻るとしよう - 予約API
  description: |
    飲食店「さて、羊に戻るとしよう」の予約管理システムAPI
    
    ## 概要
    - 顧客向け: 予約申請、残り食数確認、スケジュール確認、取引先一覧
    - 管理者向け: 予約管理、スケジュール管理、取引先管理
    
    ## 認証
    - 保護された管理者 API: `Authorization: Bearer {access_token}`（JWT、有効期限 1 時間）
    - ログイン時にアクセストークンを JSON で返し、リフレッシュトークンは **HttpOnly Cookie**（`refresh_token`）で返す。JSON にリフレッシュトークンは含めない
    - ページ再読み込み時は `POST /admin/refresh`（`credentials: include`）でアクセストークンを再取得する
  version: 1.2.0
  contact:
    name: satehits
    url: https://www.instagram.com/satehits/
servers:
  - url: http://localhost:8080/api/v1
    description: ローカル開発環境
  - url: https://api.satehits.com/api/v1
    description: 本番環境
...
paths:
...
  /reservations:
    post:
      tags:
        - 顧客向け
      summary: 予約を申請
      description: |
        予約を申請します。申請後はオーナーの承認が必要です。
        
        ### バリデーションルール
        - 人数: 1〜7名
        - 来店日: 翌日〜14日後まで
        - 来店日: その日の営業設定（有効なスケジュール）において予約不可の日は不可（`daily_schedules` を正、行が無い日はドメイン既定で解決）
        - 来店時間: 営業時間内
        
        ### 注意事項
        - 8名以上はInstagramへ
        - 当日予約はInstagramへ
        - 2週間以上先はInstagramへ
        - イベント日・特別メニュー日は注意書きが表示されます
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: 予約申請成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "409":
          description: |
            - `CAPACITY_EXCEEDED`: 予約可能数超過
            - `RESERVATION_CONFLICT`: 同一来店日時・電話番号のアクティブ予約が既にある
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /schedules:
    get:
      tags:
        - 顧客向け
      summary: 月間スケジュールを取得
      description: 指定月のスケジュール一覧を取得します。カレンダー表示用。
      operationId: getMonthlySchedule
      parameters:
        - name: year
          in: query
          required: true
          description: 年
          schema:
            type: integer
            example: 2026
        - name: month
          in: query
          required: true
          description: 月（1-12）
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 2
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyScheduleResponse"
...
  /admin/login:
    post:
      tags:
        - 認証
      summary: ログイン
      description: |
        メールアドレスとパスワードで認証し、アクセストークン（JWT）を JSON で返します。
        リフレッシュトークンはレスポンスボディに含めず、`Set-Cookie`（`refresh_token`）で返します。

        ### リフレッシュトークン Cookie
        - `HttpOnly; Secure; SameSite=None; Path=/api/v1/admin; Max-Age=2592000`（30日・ログイン時点から。refresh 成功時も同様に再設定）
        - `Domain` はサーバの `COOKIE_DOMAIN` 環境変数（開発時は未設定可）
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功（アクセストークンは JSON、リフレッシュトークンは Cookie）
          headers:
            Set-Cookie:
              description: リフレッシュトークン（`refresh_token`）。HttpOnly
              schema:
                type: string
                example: refresh_token=abc123; HttpOnly; Secure; SameSite=None; Path=/api/v1/admin; Max-Age=2592000
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refresh:
    post:
      tags:
        - 認証
      summary: アクセストークンを更新
      description: |
        Cookie のリフレッシュトークンで新しいアクセストークンを発行します。
        成功時はリフレッシュトークンをローテーションし、新しい値を `Set-Cookie` で返します。
        新 RT の有効期限は **発行時刻から30日**（スライディング）。30日以内に再度 refresh されれば期限は延びる。

        ### クライアント
        - ボディは不要
        - `credentials: include` で呼び出すこと（クロスオリジン時）
        - フロントは取得したアクセストークンをメモリに保持する（`localStorage` 非推奨）

        ### CSRF 緩和
        `Origin` または `Referer` がサーバの許可オリジン（`CORS_ORIGINS`）と一致すること。
      operationId: refreshToken
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          description: ログイン時に付与されたリフレッシュトークン
          schema:
            type: string
      responses:
        "200":
          description: 更新成功
          headers:
            Set-Cookie:
              description: ローテーション後のリフレッシュトークン
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: リフレッシュトークン無効・期限切れ・再利用検知
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidToken:
                  value:
                    error:
                      code: INVALID_TOKEN
                      message: セッションが無効です。再度ログインしてください
        "403":
          description: Origin / Referer が許可リスト外
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/logout:
    post:
      tags:
        - 認証
      summary: ログアウト
      description: |
        リフレッシュトークンをサーバ側で revoke し、Cookie を削除します。
        `Authorization: Bearer` に有効なアクセストークンを付与してください。
        `Origin` / `Referer` は `refresh` と同様に検証します。
      operationId: logout
      security:
        - BearerAuth: []
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema:
            type: string
      responses:
        "204":
          description: ログアウト成功
          headers:
            Set-Cookie:
              description: リフレッシュトークン削除（Max-Age=0）
              schema:
                type: string
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Origin / Referer が許可リスト外
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========================================
  # 予約管理API（管理者向け）
  # ========================================
  /admin/reservations:
    get:
      tags:
        - 予約管理
      summary: 予約一覧を取得
      description: 日付やステータスで絞り込んで予約一覧を取得します。
      operationId: getReservations
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: query
          required: false
          description: 日付で絞り込み（YYYY-MM-DD形式）
          schema:
            type: string
            format: date
        - name: status
          in: query
          required: false
          description: ステータスで絞り込み
          schema:
            $ref: "#/components/schemas/ReservationStatus"
        - name: source
          in: query
          required: false
          description: 予約経路で絞り込み
          schema:
            $ref: "#/components/schemas/ReservationSource"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationListResponse"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - 予約管理
      summary: 予約を登録（オーナー用）
      description: |
        Instagram/電話/知人経由などの予約をオーナーが登録します。
        Webからの予約と異なり、ステータスを指定可能です。
      operationId: createReservationByAdmin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationByAdminRequest"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 同一来店日時・電話番号のアクティブ予約が既にある（`RESERVATION_CONFLICT`）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/schedules:
    get:
      tags:
        - スケジュール管理
      summary: スケジュール一覧を取得
      description: 指定期間のスケジュール一覧を取得します。
      operationId: getSchedules
      security:
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 2
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleListResponse"
        "400":
          description: |
            クエリ `year` / `month` の欠落・非数値、または範囲外（年 2000〜2100、月 1〜12）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
              examples:
                missingQuery:
                  summary: year / month 未指定
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: 入力内容に誤りがあります
                      details:
                        - field: year
                          message: 年は必須です
                        - field: month
                          message: 月は必須です
                invalidYear:
                  summary: year が数値でない
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: 入力内容に誤りがあります
                      details:
                        - field: year
                          message: 年は数値で指定してください
                yearOutOfRange:
                  summary: year が範囲外
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: 入力内容に誤りがあります
                      details:
                        - field: year
                          message: 年は2000〜2100の範囲で指定してください
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/schedules/{date}:
    get:
      tags:
        - スケジュール管理
      summary: 日別スケジュールを取得
      operationId: getSchedule
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - スケジュール管理
      summary: 日別スケジュールを設定
      description: |
        指定日のスケジュールを設定します（Upsert）。
        パス `{date}` で日付を指定し、リクエストボディに `date` は含めません。
        初回作成・更新のいずれも 200 OK を返します。
      operationId: setSchedule
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetScheduleRequest"
      responses:
        "200":
          description: 設定成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleResponse"
        "400":
          description: リクエスト不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - スケジュール管理
      summary: 日別スケジュールを削除
      description: 指定日のスケジュールを削除し、デフォルト設定に戻します。
      operationId: deleteSchedule
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        "204":
          description: 削除成功
        "401":
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ========================================
  # 取引先管理API（管理者向け）
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReservationStatus:
      type: string
      enum: [pending, approved, rejected, cancelled, no_show]
      description: |
        予約ステータス
        - `pending`: 申請中
        - `approved`: 承認済み
        - `rejected`: 拒否
        - `cancelled`: キャンセル（メール連絡後、オーナーが手動更新）
        - `no_show`: 当日来店なし
    ReservationSource:
      type: string
      enum: [web, instagram, phone, walk_in, other]
      description: |
        予約経路
        - `web`: Webサイトからの予約
        - `instagram`: Instagram DMからの予約
        - `phone`: 電話からの予約
        - `walk_in`: 飛び込み・直接来店
        - `other`: その他（知人など）
    ScheduleType:
      type: string
      enum: [normal, morning, event, special_menu, closed]
      description: |
        スケジュールタイプ
        - `normal`: 通常営業（月火水）
        - `morning`: 朝営業（土日祝）
        - `event`: 店内イベント等（`event_name` / `event_description` 必須。営業時刻省略時は曜日別デフォルト）
        - `special_menu`: 特別メニュー
        - `closed`: 臨時休業（営業時刻は保存しない）

    # ========================================
    # 共通
    # ========================================
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # ========================================
    # 顧客向け
    # ========================================
...
    CreateReservationRequest:
      type: object
      required: [name, people, visit_date, visit_time, phone, email, recaptcha_token]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        people:
          type: integer
          minimum: 1
          maximum: 7
        visit_date:
          type: string
          format: date
        visit_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
        phone:
          type: string
        email:
          type: string
          format: email
        note:
          type: string
          maxLength: 500
        recaptcha_token:
          type: string
    ReservationResponse:
      type: object
      required: [id, name, people, visit_date, visit_time, phone, email, status, source, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        people:
          type: integer
        visit_date:
          type: string
          format: date
        visit_time:
          type: string
        phone:
          type: string
        email:
          type: string
        note:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ReservationStatus"
        source:
          $ref: "#/components/schemas/ReservationSource"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MonthlyScheduleResponse:
      type: object
      required: [year, month, schedules]
      properties:
        year:
          type: integer
        month:
          type: integer
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/DaySchedule"
    DaySchedule:
      type: object
      required: [date, schedule_type, capacity, available, is_holiday]
      properties:
        date:
          type: string
          format: date
        schedule_type:
          $ref: "#/components/schemas/ScheduleType"
        capacity:
          type: integer
        available:
          type: integer
        event_name:
          type: string
          nullable: true
        event_description:
          type: string
          nullable: true
        is_holiday:
          type: boolean
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [token, expires_at, user]
      description: |
        アクセストークン（JWT）とユーザー情報。
        リフレッシュトークンは JSON に含めず、レスポンスの `Set-Cookie`（`refresh_token`）のみで渡す。
      properties:
        token:
          type: string
          description: アクセストークン（JWT、HS256、有効期限 1 時間）
        expires_at:
          type: string
          format: date-time
          description: アクセストークンの有効期限（ISO 8601）
        user:
          type: object
          required: [id, email, role]
          properties:
            id:
              type: integer
            email:
              type: string
            role:
              type: string
              enum: [owner, developer]
    RefreshResponse:
      type: object
      required: [token, expires_at]
      description: |
        リフレッシュ成功時のアクセストークン。
        新しいリフレッシュトークンは `Set-Cookie` で返す（JSON には含めない）。
      properties:
        token:
          type: string
          description: 新しいアクセストークン（JWT）
        expires_at:
          type: string
          format: date-time
        user:
          type: object
          description: 省略可。含む場合は LoginResponse の user と同型
          properties:
            id:
              type: integer
            email:
              type: string
            role:
              type: string
              enum: [owner, developer]

    # ========================================
    # 予約管理
    # ========================================
    ReservationListResponse:
      type: object
      required: [reservations, total]
      properties:
        reservations:
          type: array
          items:
            $ref: "#/components/schemas/ReservationResponse"
        total:
          type: integer
    CreateReservationByAdminRequest:
      type: object
      required: [name, people, visit_date, visit_time, phone, email, source]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        people:
          type: integer
          minimum: 1
          maximum: 7
        visit_date:
          type: string
          format: date
        visit_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
        phone:
          type: string
        email:
          type: string
          format: email
        note:
          type: string
          maxLength: 500
        source:
          $ref: "#/components/schemas/ReservationSource"
        status:
          $ref: "#/components/schemas/ReservationStatus"
          description: 省略時はapproved
...
    ScheduleListResponse:
      type: object
      required: [year, month, schedules]
      properties:
        year:
          type: integer
        month:
          type: integer
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleResponse"
    ScheduleResponse:
      type: object
      required: [date, schedule_type, capacity]
      properties:
        date:
          type: string
          format: date
        schedule_type:
          $ref: "#/components/schemas/ScheduleType"
        capacity:
          type: integer
        event_name:
          type: string
          nullable: true
        event_description:
          type: string
          nullable: true
        open_time:
          type: string
          nullable: true
        last_order_time:
          type: string
          nullable: true
        close_time:
          type: string
          nullable: true
        is_default:
          type: boolean
          description: デフォルト設定かどうか
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SetScheduleRequest:
      type: object
      required: [schedule_type]
      properties:
        schedule_type:
          $ref: "#/components/schemas/ScheduleType"
        capacity:
          type: integer
          minimum: 0
          description: 省略時はデフォルト（10）
        event_name:
          type: string
          maxLength: 100
          description: |
            `schedule_type` が `event` のとき必須。`special_menu` ではメニュー名として任意。
        event_description:
          type: string
          maxLength: 500
          description: |
            `schedule_type` が `event` のとき必須（顧客に表示）。`special_menu` では任意。
        open_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: |
            `normal` / `morning` / `special_menu` で省略時は店舗デフォルトを適用。
            `event` で省略時は曜日別デフォルト（平日=通常、土日=朝）。`closed` では無視（常に NULL）。
        last_order_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: open_time と同様の適用ルール
        close_time:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: open_time と同様の適用ルール

    # ========================================
    # 取引先管理
    # ========================================
...